suppress arm 64 compiling warning

+++ LDProgressView/UIColor+RGBValues.m
@@ -41,7 +41,7 @@
 - (UIColor *)lighterColor {
     if ([self isEqual:[UIColor whiteColor]]) return [UIColor colorWithWhite:0.99 alpha:1.0];
     if ([self isEqual:[UIColor blackColor]]) return [UIColor colorWithWhite:0.01 alpha:1.0];
-    float hue, saturation, brightness, alpha, white;
+    CGFloat hue, saturation, brightness, alpha, white;
     if ([self getHue:&hue saturation:&saturation brightness:&brightness alpha:&alpha]) {
         return [UIColor colorWithHue:hue
                           saturation:saturation
@@ -56,7 +56,7 @@
 - (UIColor *)darkerColor {
     if ([self isEqual:[UIColor whiteColor]]) return [UIColor colorWithWhite:0.99 alpha:1.0];
     if ([self isEqual:[UIColor blackColor]]) return [UIColor colorWithWhite:0.01 alpha:1.0];
-    float hue, saturation, brightness, alpha, white;
+    CGFloat hue, saturation, brightness, alpha, white;
     if ([self getHue:&hue saturation:&saturation brightness:&brightness alpha:&alpha]) {
         return [UIColor colorWithHue:hue
                           saturation:saturation
